+++ src/main/resources/email/doc/invite_template.html
@@ -1,0 +1,25 @@
+<h3>안녕하세요, {{수신자명}}님.</h3>
+
+<p><b>{{사업자명}}</b>에서 운영 중인 판매자 <b>{{초대한사람명}}</b>님께서
+    DESKIT 플랫폼에서 함께 활동할 동업자(직원)로 {{수신자명}}님을 초대하셨습니다.</p>
+
+<p>아래 버튼을 클릭하시면 초대 수락 후 회원가입을 진행하실 수 있어요.</p>
+
+<p>👉 초대 수락하기: {{초대링크}}</p>
+
+━━━━━━━━━━━━━━━━━━━━━━━
+<h3>✔ 초대 정보</h3>
+
+<p>사업자명: <b>{{사업자명}}</b></p>
+<p>초대한 사람: <b>{{초대한사람명}}</b></p>
+<p>부여될 권한: <b>판매자 매니저(MANAGER)</b></p>
+<p>초대 만료일: <b>{{만료일}}</b></p>
+
+<p>※ 초대 링크는 만료일 이후 사용할 수 없습니다.<br>※ 본인이 요청한 초대가 아니라면 이 메일을 무시해 주세요.</p>
+━━━━━━━━━━━━━━━━━━━━━━━
+
+<p>DESKIT은 판매자와 파트너가 함께 성장하는 플랫폼입니다.<br>
+궁금한 점이 있다면 언제든지 고객센터로 문의해 주세요.
+</p>
+<p>감사합니다.<br>DESKIT 팀 드림
+</p>
